pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bestie-bench"
version = "0.1.0"
description = "Benchmarking harness for Bestie AI agent — tool calling, advice, empathy"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepeval>=0.21.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "rich>=13.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
bestie-bench = "bestie_bench.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
